requirements.txt
appnope==0.1.0
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.1
cycler==0.10.0
dateparser==0.7.6
decorator==4.4.2
deprecation==2.1.0
docopt==0.6.2
firebase-admin==4.4.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
future==0.17.1
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
graphviz==0.13.2
grpcio==1.32.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
more-itertools==8.2.0
msgpack==1.0.0
nltk==3.5
np==1.0.2
num2words==0.5.10
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.3
pandas==1.0.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.13.0
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.8.1
pyaml==19.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclustering==0.9.3.1
pycparser==2.20
pydotplus==2.0.2
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-twitter==3.5
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sklearn-crfsuite==0.3.6
snips-nlu==0.20.2
snips-nlu-parsers==0.4.3
snips-nlu-utils==0.9.1
SQLAlchemy==1.3.16
tabulate==0.8.7
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.1.0
